refactor: db 관련 책임을 dbHelper로 이동
